Доработка sprint_11 после второго  Code review
