+++ ImageFeed/ImagesListView/ImagesListCell.swift
@@ -1,4 +1,3 @@
-import Foundation
 import UIKit
 
 final class ImagesListCell: UITableViewCell {
