+++ ImageFeed/ProgressHUD/ProgressHUD.swift
@@ -1,0 +1,20 @@
+import UIKit
+import ProgressHUD
+
+//MARK: -
+
+final class UIBlockingProgressHUD {
+    private static var window: UIWindow? {
+        return UIApplication.shared.windows.first
+    }
+    
+    static func show() {
+        window?.isUserInteractionEnabled = false
+        ProgressHUD.show()
+    }
+    
+    static func dismiss() {
+        window?.isUserInteractionEnabled = true
+        ProgressHUD.dismiss()
+    }
+}
