+++ ImageFeed/Services/OAuth2TokenStorage.swift
@@ -1,4 +1,3 @@
-import Foundation
 import SwiftKeychainWrapper
 import UIKit
 
@@ -6,7 +5,7 @@ import UIKit
 final class OAuth2TokenStorage {
     
     static let shared = OAuth2TokenStorage()
-    private init() {print("Синглтон инициализирован")}
+    private init() {}
     private let keyChain = KeychainWrapper.standard
     private let bearerToken = "bearerToken"
