+++ ImageFeed/TabBarController/TabBarController.swift
@@ -1,6 +1,4 @@
-import Foundation
 import UIKit
-
 
 final class TabBarViewController: UITabBarController {
